needed to move transformSkillRatingsForDB back to main component because of UseAuth hook.
